Move overlay fille
